project/requirements.txt
python-dotenv
aiofiles
fastapi
uvicorn
pydantic
Jinja2
pytest
celery
flower
redis
rq
pymongo
requests
httpx
pika
rabbitmq_worker
aio_pika
msal
python-multipart
sentry-sdk[fastapi]
patool
boto3